converter/requirements.txt
# DICOM WSI Converter Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# DICOM/WSI processing
wsidicom>=0.20.0
wsidicomizer>=0.24.0
pydicom>=2.4.0
highdicom>=0.22.0
openslide-python>=1.3.0

# Philips iSyntax support (open-source - install from GitHub for Python 3.11+ support)
# Install pyisyntax from GitHub (required for iSyntax support in wsidicomizer)
pyisyntax @ git+https://github.com/anibali/pyisyntax.git@master

# Image processing
Pillow>=10.0.0
numpy>=1.26.0
imagecodecs>=2024.1.0
# For JPEG2000 support in Pillow
pillow-heif>=0.15.0

# HTTP client for Orthanc
httpx>=0.26.0
requests>=2.31.0

# Task queue (for async conversions)
celery>=5.3.0
redis>=5.0.0

# Utilities
watchdog>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Monitoring
prometheus-client>=0.19.0

